Add 'merge_hashes' option to include_vars module
This commit adds 'merge_hashes' option to 'include_vars' module which
makes a new hash variable merge with the existing hash variable of the
same name instead of overwriting it.

This is useful in cases where you organize configuration files in
a certain hierarchy and want nested variables to recursively merge
without having to use the global 'hash_behaviour = merge' setting which
is now deprecated.